refactor: rename TransactionsView to HistoryView.kt
